requirements.txt
adal==1.2.7
APScheduler==3.6.3
asgiref==3.5.2
azure-core==1.24.0
azure-identity==1.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-grappelli==3.0.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
gspread==5.3.2
httplib2==0.20.4
idna==3.3
msal==1.17.0
msal-extensions==1.0.0
msgraph-core==0.2.2
mysqlclient==2.1.0
oauth2client==4.1.3
oauthlib==3.2.0
portalocker==2.4.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-msgraph==0.2.8
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
